test: add yaml templates for aidbox and fhir format
