+++ ts/server/src/utils/extract.spec.ts
@@ -94,11 +94,20 @@ describe('Transformation', () => {
         ).toStrictEqual(undefined);
     });
 
+    test('for multiline template works properly', () => {
+        expect(
+            resolveTemplate(
+                resource,
+                '{{\nlist.where(\n$this.key=1\n).key\n}}',
+            ),
+        ).toStrictEqual(1);
+    });
+
     test('fails with incorrect fhirpath expression', () => {
         expect(() => resolveTemplate({} as any, "{{ item.where(linkId='a) }}")).toThrowError(
             FPMLValidationError,
         );
-    });
+    }); 
 });
 
 describe('Context block', () => {
